+++ changelog.md
@@ -1,3 +1,8 @@
 # Changelog
 
-Opened 'unit-tests' branch.
+**./submission.js**
+* Removed comments:
+	* Run example cases.
+	* Display results.
+* Removed console output for example cases.
+* Commented out example cases.